fix: 수불부 ADJUST 타입 반영 + 기타출고 kg 소수점 입력 지원
1. 수불부(ledger) ADJUST 타입 미분류 수정
   - 양수(+) 조정 → 입고 컬럼에 반영
   - 음수(-) 조정 → 출고 컬럼에 반영
   - PDF는 이미 qty 부호로 분류하여 정상 작동

2. 기타출고 소수점 수량 입력 지원
   - blueprint: int() → float() 변환
   - template: input에 step=any 추가 (기본행 + 동적행)

+++ blueprints/etc_outbound.py
@@ -103,7 +103,7 @@ def process():
     for i in range(len(product_names)):
         name = product_names[i].strip() if i < len(product_names) else ''
         try:
-            qty = int(qtys[i]) if i < len(qtys) else 0
+            qty = float(qtys[i]) if i < len(qtys) else 0
         except (ValueError, IndexError):
             qty = 0
         reason = reasons[i] if i < len(reasons) else '기타'
